New version of greedy search for prune.

+++ disk-ann/src/main/java/jetbrains/exodus/diskann/GreedyVertex.java
@@ -2,17 +2,17 @@ package jetbrains.exodus.diskann;
 
 final class GreedyVertex implements Comparable<GreedyVertex> {
     final int index;
-    double distance;
+    float distance;
 
     boolean isPqDistance;
 
-    GreedyVertex(int index, double distance, boolean isPqDistance) {
+    GreedyVertex(int index, float distance, boolean isPqDistance) {
         this.index = index;
         this.distance = distance;
         this.isPqDistance = isPqDistance;
     }
 
     public int compareTo(GreedyVertex other) {
-        return Double.compare(distance, other.distance);
+        return Float.compare(distance, other.distance);
     }
 }
